// src/interfaces/project/ProjectSenttings.ts
export interface ProjectSettings {
  restForGrowth: number;
  expectedUserSpeedGbps: number;
  horizontalCablingLengthMeters: number;
  patchCordLengthMeters: number;
  patchCableLengthMeters: number;
  pigtailReserveLengthMeters: number;
  haveCameras: boolean;
  havePhones: boolean;
  haveAccessPoints: boolean;
  haveAccessControllers: boolean;
  opticalCordLength: number;
}